+++ govtool/backend/src/VVA/Types.hs
@@ -62,6 +62,7 @@ data Vote
       , voteDocHash    :: Maybe Text
       , voteEpochNo    :: Integer
       , voteDate       :: UTCTime
+      , voteTxHash     :: Text
       }
 
 data DRepInfo
@@ -74,6 +75,7 @@ data DRepInfo
       , dRepInfoUrl                      :: Maybe Text
       , dRepInfoDataHash                 :: Maybe Text
       , dRepInfoVotingPower              :: Maybe Integer
+      , dRepInfoLatestTxHash             :: Maybe Text
       }
 
 data DRepStatus = Retired | Active | Inactive
@@ -82,14 +84,15 @@ data DRepType = DRep | SoleVoter
 
 data DRepRegistration
   = DRepRegistration
-      { dRepRegistrationDRepHash    :: Text
-      , dRepRegistrationView        :: Text
-      , dRepRegistrationUrl         :: Maybe Text
-      , dRepRegistrationDataHash    :: Maybe Text
-      , dRepRegistrationDeposit     :: Integer
-      , dRepRegistrationVotingPower :: Maybe Integer
-      , dRepRegistrationStatus      :: DRepStatus
-      , dRepRegistrationType        :: DRepType
+      { dRepRegistrationDRepHash     :: Text
+      , dRepRegistrationView         :: Text
+      , dRepRegistrationUrl          :: Maybe Text
+      , dRepRegistrationDataHash     :: Maybe Text
+      , dRepRegistrationDeposit      :: Integer
+      , dRepRegistrationVotingPower  :: Maybe Integer
+      , dRepRegistrationStatus       :: DRepStatus
+      , dRepRegistrationType         :: DRepType
+      , dRepRegistrationLatestTxHash :: Maybe Text
       }
 
 data Proposal
@@ -110,6 +113,7 @@ data Proposal
       , proposalMotivaiton     :: Maybe Text
       , proposalRationale      :: Maybe Text
       , proposalMetadata       :: Maybe Value
+      , proposalReferences     :: Maybe Value
       , proposalYesVotes       :: Integer
       , proposalNoVotes        :: Integer
       , proposalAbstainVotes   :: Integer