+++ govtool/frontend/src/components/molecules/types.ts
@@ -26,3 +26,14 @@ export type EmptyStateGovernanceActionsCategoryProps = {
   category?: string;
   isSearch?: boolean;
 };
+
+export type AutomatedVotingCardProps = {
+  description: string;
+  inProgress?: boolean;
+  isConnected?: boolean;
+  isSelected?: boolean;
+  onClickDelegate: () => void;
+  onClickInfo: () => void;
+  title: string;
+  votingPower: string | number;
+};