+++ govtool/frontend/src/consts/dRepDirectory/sorting.ts
@@ -1,0 +1,14 @@
+export const DREP_DIRECTORY_SORTING = [
+  {
+    key: "NewestRegistered",
+    label: "Newest registered",
+  },
+  {
+    key: "VotingPower",
+    label: "Voting power",
+  },
+  {
+    key: "Status",
+    label: "Status",
+  },
+];