+++ govtool/frontend/src/consts/icons.ts
@@ -15,6 +15,8 @@ export const ICONS = {
   dashboardIcon: "/icons/Dashboard.svg",
   download: "/icons/Download.svg",
   drawerIcon: "/icons/DrawerIcon.svg",
+  dRepDirectoryActiveIcon: "/icons/DRepDirectoryActive.svg",
+  dRepDirectoryIcon: "/icons/DRepDirectory.svg",
   externalLinkIcon: "/icons/ExternalLink.svg",
   faqsActiveIcon: "/icons/FaqsActive.svg",
   faqsIcon: "/icons/Faqs.svg",
@@ -32,4 +34,5 @@ export const ICONS = {
   sortWhiteIcon: "/icons/SortWhite.svg",
   timerIcon: "/icons/Timer.svg",
   warningIcon: "/icons/Warning.svg",
+  loaderIcon: "/icons/Loader.svg",
 };