+++ govtool/frontend/src/models/index.ts
@@ -1,3 +1,4 @@
 export * from "./api";
 export * from "./snackbar";
 export * from "./wallet";
+export * from "./metadataValidation";